Move some parts of the SupermatterIntegritySystem to Shared
